requirements.txt
streamlit>=1.20.0
langchain>=0.1.0
openai>=1.0.0
chromadb>=0.4.0
pdfplumber
pypdf
python-dotenv
langchain-community
langchain-openai
langchain-core
langchain-text-splitters